Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/text/2d.text.drawing.style.spacing.worker-expected.txt


FAIL Testing letter spacing and word spacing assert_equals: ctx.textLetterSpacing === 0 (got [undefined], expected 0[number]) expected (number) 0 but got (undefined) undefined
